Realizamos mejoras en nuestro codigo, aplicamos herencia (ClasePrincipal) de donde herendan (Propiedad y Vendedor) la cual nos permite reutilizar mucho codigo como por ejemplo la conexion a base de datos se hace todo en la clase principal, o los metodos comunes a las clases hijas, etc.
