+++ clases/Vendedor.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App;
+
+class Vendedor extends ClasePrincipal {
+/*cuando se llame al metodo all desde un objeto vendedor el atributo $tabla va a tomar el valor de 'vendedores'*/
+    protected static $tabla = 'vendedores';
+    protected static $columnasDB = ['id', 'nombre', 'apellido', 'telefono'];
+
+    public $id;
+    public $nombre;
+    public $apellido;
+    public $telefono;
+
+    public function __construct($args = [])
+    {
+        $this->id = $args['id'] ?? null;
+        $this->nombre = $args['nombre'] ?? '';
+        $this->apellido = $args['apellido'] ?? '';
+        $this->telefono = $args['telefono'] ?? '';
+    }
+}
